Created an example of cool uses
This example will post information like:
when a player enters a cave layer, it shows on your screen what number layer you're on. Also sends message to a discord channel.
when a player dies it sends a message to a discord channel
when a player joins or leaves sends a message to a discord channel

It also automatically accepts invites it receives while running.