Major changes (click to see details)
Major changes in this version:
•create_console now opens a connection to a server's console not only once, but now it will open the console connection automatically when your server starts up.
So all that you need to do is use create_console once, and it'll open the server console connection the moment the server starts forever.
•console_sub now does something similar. Now console_sub will subscribe to the subscription every time the server starts up. So you only have to console_sub once.

Before this version, if you created console you'd have to make sure the server was up first, and when the server went down you'd lose your subscriptions and it would not connect when the server started back up. Now everything is handled for you. This brings the program's stability to near 100% (only near since we've got some stress testing to do! otherwise it's been perfect.)