Changed the variable self.main_subscriptions data.
The variable now stores a dict with a value that contains another dict, rather than a variable with a single key and value.
This was done to create a way to ensure stability on the main_websocket. Before, if Alta restarted their websocket, the subscriptions to the websocket would be lost. Now if their end crashes, it still retains the subscriptions in py_tale's code, and re-subs to them.