Updated create_console and added custom Exceptions
Added custom Exceptions for usage in try: except: statements to allow more control over errors.

A majorly needed change was added in this commit. create_console will now wait for the server console connection to be successfully opened before continuing. This makes it possible to ensure you're setup before trying to do things with the server. Upon failing to open the server, one of a few custom exceptions will be thrown.

create_console now sports a new optional parameter in this version `timeout` which is the max time in seconds the program should try to open the console before throwing an exception.

Also some small changes to the program's structure.