Updated create_console and added new Exception
Added a new Exception for when you lack permissions to get info about/use a console.

Made the create_console function take an optional parameter that allows you to wait for the console connection to be fully open before continuing. This is an optional parameter that was added for dual functionality.

This allows you to ensure the console is open before continuing, or allows you to not care if it is or not and continue anyways. The console will always open the connection when the server starts up anyways.