Added custom exception and changed create_console
create_console now works as I imagined it to. Leaving create_console with just the server_id argument will allow you to just let the program handle opening it whenever it goes online. ensure_open and timeout is a way we can make the function throw an error if the server is offline. Basically allows us to ensure we've got the console open before doing stuff.