Fixed another rare migration bug from Alta's end.
After ~50 days consecutively of running a bot with py-tale, it crashed with this error:

`{'id': 26, 'event': 'response', 'key': 'GET /ws/migrate', 'content': '{"message":"IDX20803: Unable to obtain configuration from: \'System.String\'.","error_code":"None"}', 'responseCode': 403}`

I believe this is either due to a bug or crash on Alta's end. In this new version of py-tale I've created a generic check to ensure the response code is `200` and if not, perform a manual migration just like it would if the response took too long from the last update of py-tale.